requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
anthropic>=0.20.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
aiortc>=1.6.0
av>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
aiohttp>=3.9.0
